backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
distro==1.9.0
dotenv==0.9.9
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.21.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
keyboard==0.13.5
MarkupSafe==3.0.2
multidict==6.5.0
numpy==2.2.6
openai==0.28.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
pypiwin32==223
pyserial==3.5
python-dotenv==1.1.0
pyttsx3==2.98
pywin32==310
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
